fix: resolve error after rebase
