+++ src/rubrix/utils/__init__.py
@@ -12,4 +12,5 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+from .span_utils import SpanUtils
 from .utils import LazyRubrixModule, limit_value_length, setup_loop_in_thread
